+++ autosubmit/platforms/wrappers/wrapper_factory.py
@@ -20,7 +20,9 @@ import re
 from autosubmit.platforms.wrappers.wrapper_builder import (
     WrapperDirector, PythonVerticalWrapperBuilder, PythonHorizontalWrapperBuilder,
     PythonHorizontalVerticalWrapperBuilder, PythonVerticalHorizontalWrapperBuilder, BashHorizontalWrapperBuilder,
-    BashVerticalWrapperBuilder, SrunHorizontalWrapperBuilder, SrunVerticalHorizontalWrapperBuilder
+    BashVerticalWrapperBuilder, SrunHorizontalWrapperBuilder, SrunVerticalHorizontalWrapperBuilder,
+    FluxHorizontalWrapperBuilder, FluxVerticalWrapperBuilder, FluxHorizontalVerticalWrapperBuilder,
+    FluxVerticalHorizontalWrapperBuilder
 )
 
 
@@ -53,6 +55,15 @@ class WrapperFactory(object):
             kwargs['queue'] = self.queue(wrapper_data.queue)
             kwargs['threads'] = self.threads(str(wrapper_data.threads))
             kwargs['reservation'] = self.reservation(wrapper_data.reservation)
+
+            # When wrapping with Flux, give it control over all resources
+            if wrapper_data.method.lower() == "flux":
+                if wrapper_data.nodes == '':
+                    kwargs['threads'] = self.threads(str(int(wrapper_data.threads) * int(kwargs['num_processors_value'])))
+                else:
+                    kwargs['threads'] = self.threads('')
+                kwargs['num_processors'] = self.processors('')
+                kwargs['tasks'] = self.tasks('')
 
         kwargs["executable"] = wrapper_data.executable
         kwargs['header_directive'] = self.header_directives(**kwargs)
@@ -165,21 +176,31 @@ class WrapperFactory(object):
 class SlurmWrapperFactory(WrapperFactory):
 
     def vertical_wrapper(self, **kwargs):
-        return PythonVerticalWrapperBuilder(**kwargs)
+        if kwargs["method"] == 'flux':
+            return FluxVerticalWrapperBuilder(**kwargs)
+        else:
+            return PythonVerticalWrapperBuilder(**kwargs)
 
     def horizontal_wrapper(self, **kwargs):
 
         if kwargs["method"] == 'srun':
             return SrunHorizontalWrapperBuilder(**kwargs)
+        elif kwargs["method"] == 'flux':
+            return FluxHorizontalWrapperBuilder(**kwargs)
         else:
             return PythonHorizontalWrapperBuilder(**kwargs)
 
     def hybrid_wrapper_horizontal_vertical(self, **kwargs):
-        return PythonHorizontalVerticalWrapperBuilder(**kwargs)
+        if kwargs["method"] == 'flux':
+            return FluxHorizontalVerticalWrapperBuilder(**kwargs)
+        else:
+            return PythonHorizontalVerticalWrapperBuilder(**kwargs)
 
     def hybrid_wrapper_vertical_horizontal(self, **kwargs):
         if kwargs["method"] == 'srun':
             return SrunVerticalHorizontalWrapperBuilder(**kwargs)
+        elif kwargs["method"] == 'flux':
+            return FluxVerticalHorizontalWrapperBuilder(**kwargs)
         else:
             return PythonVerticalHorizontalWrapperBuilder(**kwargs)
 
@@ -220,21 +241,31 @@ class SlurmWrapperFactory(WrapperFactory):
 class PJMWrapperFactory(WrapperFactory):
 
     def vertical_wrapper(self, **kwargs):
-        return PythonVerticalWrapperBuilder(**kwargs)
+        if kwargs["method"] == 'flux':
+            raise NotImplementedError(self.exception)   # pragma: no cover
+        else:
+            return PythonVerticalWrapperBuilder(**kwargs)
 
     def horizontal_wrapper(self, **kwargs):
 
         if kwargs["method"] == 'srun':
             return SrunHorizontalWrapperBuilder(**kwargs)
+        elif kwargs["method"] == 'flux':
+            raise NotImplementedError(self.exception)   # pragma: no cover
         else:
             return PythonHorizontalWrapperBuilder(**kwargs)
 
     def hybrid_wrapper_horizontal_vertical(self, **kwargs):
-        return PythonHorizontalVerticalWrapperBuilder(**kwargs)
+        if kwargs["method"] == 'flux':
+            raise NotImplementedError(self.exception)   # pragma: no cover
+        else:
+            return PythonHorizontalVerticalWrapperBuilder(**kwargs)
 
     def hybrid_wrapper_vertical_horizontal(self, **kwargs):
         if kwargs["method"] == 'srun':
             return SrunVerticalHorizontalWrapperBuilder(**kwargs)
+        elif kwargs["method"] == 'flux':
+            raise NotImplementedError(self.exception)   # pragma: no cover
         else:
             return PythonVerticalHorizontalWrapperBuilder(**kwargs)
 
@@ -270,10 +301,16 @@ class PJMWrapperFactory(WrapperFactory):
 class EcWrapperFactory(WrapperFactory):
 
     def vertical_wrapper(self, **kwargs):
-        return BashVerticalWrapperBuilder(**kwargs)
+        if kwargs["method"] == 'flux':
+            raise NotImplementedError(self.exception)   # pragma: no cover
+        else:
+            return BashVerticalWrapperBuilder(**kwargs)
 
     def horizontal_wrapper(self, **kwargs):
-        return BashHorizontalWrapperBuilder(**kwargs)
+        if kwargs["method"] == 'flux':
+            raise NotImplementedError(self.exception)   # pragma: no cover
+        else:
+            return BashHorizontalWrapperBuilder(**kwargs)
 
     def header_directives(self, **kwargs):
         return self.platform.wrapper_header(**kwargs)